Renamed special_wrappers.{c,h} -> xsf_wrappers.{c,h}
