+++ scipy/special/_agm.pxd
@@ -5,7 +5,7 @@ import cython
 from libc.math cimport log, exp, fabs, sqrt, isnan, isinf, NAN, M_PI
 
 
-cdef extern from "special_wrappers.h" nogil:
+cdef extern from "xsf_wrappers.h" nogil:
     double cephes_ellpk_wrap(double x)
