+++ scipy/special/_ellipk.pxd
@@ -1,4 +1,4 @@
-cdef extern from "special_wrappers.h" nogil:
+cdef extern from "xsf_wrappers.h" nogil:
     double cephes_ellpk_wrap(double x)
