+++ scipy/special/_hypergeometric.pxd
@@ -4,7 +4,7 @@ import cython
 
 from . cimport sf_error
 
-cdef extern from 'special_wrappers.h':
+cdef extern from 'xsf_wrappers.h':
     double hypU_wrap(double, double, double) nogil
     double cephes_poch_wrap(double x, double m) nogil
